(PC-17012)[API] refactor: rework public collective offer api
